backend/requirements.txt
langchain
langchain-openai
langchain-experimental
flask
redis
flask-cors
pyjwt
schedule
pymysql
python-dotenv
cryptography